wip part 2

+++ Day19/Day19.h
@@ -10,7 +10,7 @@ class Day19 {
 public:
     static int Part1();
 
-    static long long Part2();
+    static uint64_t Part2();
 
     static bool isPossible(const vector<string> *towels, const string &pattern);
